app/py-agent/requirements.txt
langgraph
openai
psycopg2-binary
python-dotenv 